Vlakke punten gefixed... Gebruikte / bij ints -> moet bij doubles! + andere graph systeem... maar scheelt volgens mij nog altijd iets aan als ik vergelijk met ppts

+++ schedulingAlg/src/Algoritmen.java
@@ -48,7 +48,7 @@ public class Algoritmen {
 			tijd += bedieningsTijd;
 			huidigProces.setEndtime(tijd);
 			huidigProces.setRuntime(tijd - aankomstTijd);
-			huidigProces.setNorRuntime((tijd - aankomstTijd) / bedieningsTijd);
+			huidigProces.setNorRuntime(((double)(tijd - aankomstTijd)) / bedieningsTijd);
 			huidigProces.setWaittime(tijd - aankomstTijd - bedieningsTijd);
 			loper++;
 		}
@@ -102,7 +102,7 @@ public class Algoritmen {
 			tijd += bedieningsTijd;
 			huidigProces.setEndtime(tijd);
 			huidigProces.setRuntime(tijd - aankomstTijd);
-			huidigProces.setNorRuntime((tijd - aankomstTijd) / bedieningsTijd);
+			huidigProces.setNorRuntime(((double)(tijd - aankomstTijd)) / bedieningsTijd);
 			huidigProces.setWaittime(tijd - aankomstTijd - bedieningsTijd);
 			loper++;
 		}
